Add attention mask to local_forward
TransformerEncoder mask is specified for the task of next token prediction, preventing forward-looking.

The mask is an upper-triangular -inf matrix of [sequence_len, sequence_len], which is standard and added to the attention matrix before SoftMax effectively removing values contributed by the token to be predicted and all tokens that come after it in the sequence.