Require sentencepiece for fast tokenization
Only fast tokenizers can produce token offsets, so fast tokenizer requirements need to be met. In particular, sentencepiece is required to address the following error:

ValueError: Couldn't instantiate the backend tokenizer from one of:
(1) a `tokenizers` library serialization file,
(2) a slow tokenizer instance to convert or
(3) an equivalent slow tokenizer class to instantiate and convert.
You need to have sentencepiece installed to convert a slow tokenizer to a fast one.