Remove bittensor tokenizer modifications
The GPT2Tokenizer defaults should be retained as much as possible, since models trained on this tokenizer fails when they encounter new special tokens that weren't in the original tokenizer.

Currently, the bittensor tokenizer is not equivalent to the GPT2Tokenizer, so logit translation may be required in cases where new special tokens are encountered. More models can be supported without translation if the original tokenizer is adhered to directly.