Skip server model run and raise FileNotFoundError
If encodings_cache_file is not available, skip the server model run and directly raise FileNotFoundError instead.

The test should just fail if we don't have cached embeddings, especially since they're in VCS.

When we run in circleCI we don't have the resources to compute these, and it leaves the test just hanging silently as it computes the embeddings. I think we should have this code in a separate script if we ever want to change what we're doing in this test.