Consume validator nucleus UID queue fully
Since epochs are now concluded at blocks_per_epoch many validators will have unqueried UIDs still in queue. It is better that these are retained and consumed in the next epoch, since this will ensure all UIDs are queried over a cycle that may involve successive epochs.

The inefficiency introduced is that if the queue has to be refilled in the same epoch, then a UID may be queried twice, although this is preferred over the risk of not querying a UID at all if the remaining queue is not fully consumed even if another epoch is required.