Update reward.py
If loss is 0, and the c_truncated_reward + r_truncated_reward = 0, then a validator with a score of 0 (those with a "Sorry, I cannot do that Dave" response) won't not get punished for sending a bad reply. This makes it so that the resulting score always needs to be positive.