Changes cacheing for get_decoder_class
Changes the cacheing for the previously implemented get_decoder_class. While it was probably safe to do this, there are a handful of edge case scenarios in the overall configuration we may have overlooked. By changing the cached method convert_type_string, we avoid the potential edge cases that may result in serious issues down the line.