test_axon.py: test whether miner runs, show failure
Issues in miner.py largely went unnoticed. This patch addresses this by
checking whether the miner indeed runs, and at the end of the test
terminating it and examining its output. It is asserted that the string
"Exception" is not present in stdout/stderr.

The subprocess is started with asyncio.create_subprocess_exec() so that
it can be terminated easily. Added benefit is that escaping and quoting
is not necessary anymore, as the arguments are passed as a list.