RFC: runtime_browser
This code exposes runtime storage items, constants and APIs as if they
are mirrored in Python. It is intended to simplify and reduce code, and
to close the current gap between subtensor code and bittensor code. As
it is based on metadata, it exposes past, current and future runtime
items.

E.g. the subnet owner map can be queried using only:

    import bittensor as bt
    bt.runtime.SubtensorModule.SubnetOwner[1]

test_runtime_browser.py contains various examples that demonstrate the
proposed concept.

The impact on bittensor is negligible, as it is implemented as a fully
separate module.